Dividing operations on messages to different traits
